Move auth folder to shared folder
