+++ ui/login-callback/index.jsx
@@ -3,8 +3,8 @@ import { render } from 'react-dom';
 import { AppContainer } from 'react-hot-loader';
 import 'font-awesome/css/font-awesome.css';
 
-import LoginCallback from './js/auth/LoginCallback';
-import './css/login.css';
+import LoginCallback from './LoginCallback';
+import '../css/login.css';
 
 const load = () => render((
   <AppContainer>
@@ -13,7 +13,7 @@ const load = () => render((
 ), document.getElementById('root'));
 
 if (module.hot) {
-  module.hot.accept('./js/auth/LoginCallback', load);
+  module.hot.accept('./LoginCallback', load);
 }
 
 load();
