+++ ui/shared/auth/AuthService.js
@@ -1,4 +1,4 @@
-import { userSessionFromAuthResult, renew, loggedOutUser } from './auth-utils';
+import { userSessionFromAuthResult, renew, loggedOutUser } from '../../helpers/auth';
 import taskcluster from '../../helpers/taskcluster';
 import { getApiUrl } from '../../helpers/url';
 import UserModel from '../../models/user';
