+++ ui/shared/auth/Login.jsx
@@ -2,12 +2,12 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import isEqual from 'lodash/isEqual';
 
-import AuthService from '../js/auth/AuthService';
-import { loggedOutUser } from '../js/auth/auth-utils';
-import taskcluster from '../helpers/taskcluster';
-import { getApiUrl, loginCallbackUrl } from '../helpers/url';
-import UserModel from '../models/user';
-import { withNotifications } from './context/Notifications';
+import AuthService from './AuthService';
+import { loggedOutUser } from '../../helpers/auth';
+import taskcluster from '../../helpers/taskcluster';
+import { getApiUrl, loginCallbackUrl } from '../../helpers/url';
+import UserModel from '../../models/user';
+import { withNotifications } from '../context/Notifications';
 
 /**
  * This component handles logging in to Taskcluster Authentication
